cobol: Correct diagnostic strings for 32-bit builds.
Avoid %z for printf-family.  Cast pid_t to long.  Avoid use of YYUNDEF
for old Bison versions.

	PR cobol/120621

gcc/cobol/ChangeLog:

	* genapi.cc (parser_compile_ecs): Cast argument to unsigned long.
	(parser_compile_dcls): Same.
	(parser_division): RAII.
	(inspect_tally): Cast argument to unsigned long.
	* lexio.cc (cdftext::lex_open): Cast pid_t to long.
	* parse.y: hard-code values for old versions of Bison, and message format.
	* scan_ante.h (wait_for_the_child): Cast pid_t to long.

(cherry picked from commit 007392c0f93cf46b9e87aebdd04e123e3381fc07)

+++ gcc/cobol/lexio.cc
@@ -1524,11 +1524,11 @@ cdftext::lex_open( const char filename[] ) {
     int status;
     auto kid = wait(&status);
     gcc_assert(pid == kid);
-    if( kid == -1 ) cbl_err( "failed waiting for pid %d", pid);
+    if( kid == -1 ) cbl_err( "failed waiting for pid %ld", static_cast<long>(pid));
 
     if( WIFSIGNALED(status) ) {
-      cbl_errx( "%s pid %d terminated by %s",
-           filter, kid, strsignal(WTERMSIG(status)) );
+      cbl_errx( "%s pid %ld terminated by %s",
+                filter, static_cast<long>(kid), strsignal(WTERMSIG(status)) );
     }
     if( WIFEXITED(status) ) {
       if( (status = WEXITSTATUS(status)) != 0 ) {
